fix(control-plane): cast LIMIT $2 to int64 in log GC
batchSize was passed as Go int to a Postgres LIMIT parameter.
On 64-bit builds this matches LIMIT's bigint type, but on
32-bit builds (or strict-type Postgres configurations), int4 in
LIMIT can fail at execution with a type-mismatch error that
sqlmock (used in tests) does not catch.

Cast to int64 explicitly so the binding always matches.

- int64(batchSize) in DeleteOlderThanBatched
- TestLogEntryRepository_DeleteOlderThanBatches_BoundInt64
- TestLogEntryRepository_DeleteOlderThanBatches_UsesServerNOW
  now uses int64(10_000) in WithArgs

+++ edge-control-plane/internal/repository/log_entry.go
@@ -138,7 +138,7 @@ func (r *LogEntryRepository) DeleteOlderThanBatched(
 		}
 		res, err := r.db.ExecContext(ctx,
 			`DELETE FROM logs WHERE id IN (SELECT id FROM logs WHERE ts < NOW() - make_interval(secs => $1) LIMIT $2)`,
-			retention.Seconds(), batchSize)
+			retention.Seconds(), int64(batchSize))
 		if err != nil {
 			return total, fmt.Errorf("deleting old logs (batch %d): %w", i, err)
 		}
